+++ lib/td_dd/data_structures/data_structure_versions.ex
@@ -2,11 +2,12 @@ defmodule TdDd.DataStructures.DataStructureVersions do
   @moduledoc """
   The DataStructureVersion specific context.
   """
-  import Bodyguard, only: [permit?: 4]
+  import Bodyguard, only: [permit?: 4, permit?: 3]
 
   alias TdCache.TemplateCache
   alias TdDd.DataStructures
   alias TdDd.DataStructures.Tags
+  alias TdDq.Implementations
 
   @controller_enrich_attrs [
     :data_fields,
@@ -167,7 +168,8 @@ defmodule TdDd.DataStructures.DataStructureVersions do
           permit?(DataStructures, :manage_grant_removal, claims, data_structure) and
             permit?(DataStructures, :manage_foreign_grant_removal, claims, data_structure),
         create_foreign_grant_request:
-          permit?(DataStructures, :create_foreign_grant_request, claims, data_structure)
+          permit?(DataStructures, :create_foreign_grant_request, claims, data_structure),
+        view_quality: permit?(Implementations, "view_quality", claims)
       }
 
       [